+++ x-pack/plugins/runtime_fields/public/components/runtime_field_editor_flyout_content/runtime_field_editor_flyout_content.tsx
@@ -66,7 +66,7 @@ export interface Props {
   defaultValue?: RuntimeField;
 }
 
-export const RuntimeFieldEditorFlyout = ({
+export const RuntimeFieldEditorFlyoutContent = ({
   onSave,
   onCancel,
   docLinks,
